+++ CHANGELOG.md
@@ -16,6 +16,11 @@ The **third number** is the patch version (bug fixes)
 
 - **BREAKING**: `HarborAsyncClient.update_robot_token` renamed to `HarborAsyncClient.refresh_robot_token` to better reflect the API endpoint name and purpose.
 
+
+### Fixed
+
+- Pagination failing when one or more query parameter values included a comma.
+
 ## [0.8.2](https://github.com/pederhan/harborapi/tree/harborapi-v0.8.2) - 2023-02-09
 
 ### Fixed